Create README.md
This is a SQL project that focuses on data cleaning within SQL. The data used is the Nashville Housing which consists of information about the housing including sale dates, sale as vacant, address, and more. Standardizing the Sale date was the first task completed so the column reads year, month, then day.  The next task was populating Null values in the "property address" with an actual address. This required exploring the date and noticing the "owner address" is matched with a "unique ID". In this case, the null values were replaced with the "owner address" based on the assigned "unique ID". The Address column was a bit busy, consisting of address, city and state in a single column. Therefore, resulting the analysis process is a bit more difficult. To resolve this, the Parsename function was used to delimit the address, city, and state into individual columns. To keep the Sold As Vacant column clean and straightforward, the Y and N were replaced with Yes and No. Instead of 4 values (Y, N, Yes, No), the column was reduced to 2 after cleaning to only Yes and No. Finally, to show cast how columns and tables can be deleted the DELETE statement was used to remove a table and the DROP statement was used to remove fields.